cartography/intel/gcp/requirements.txt
google-api-python-client==2.131.0
google-auth==2.35.0
neo4j==5.22.0
google-cloud-storage==2.9.0
